implement errno in kernel; adjust code structure

+++ os/src/trap/mod.rs
@@ -25,8 +25,7 @@ pub fn trap_handler() {
             context.sepc += 4;
             context.x[10] =
                 syscall(context.x[17],
-                        [context.x[10], context.x[11], context.x[12], context.x[13], context.x[14]])
-                    as usize;
+                        [context.x[10], context.x[11], context.x[12], context.x[13], context.x[14]]);
         },
         Trap::Interrupt(Interrupt::SupervisorTimer) => {
             stop_current_and_run_next_task();
